Some code refactoring and most importantly added an EEPROM to store and read the levels from in a one by one manner. (#2)
* load levels from the sdcard

many design limitations encountered. added some workarounds but a refactor of level handling needs to be put in place.
SD library used is WaveHC which I use only in read-only mode. This was the best alternative as it is the only one that fits within the SRAM and Flash bounds of the board and still allows the game to somewhat be changed to work. (somewhat is used because I hit the limits of SRAM and strange stuff is happening hence why I disabled adding enemies for now. Will refactor memory usage stuff via another brach.)
Note to self and other who stumble to this: plan and prototype your project before soldering it on board :). Don't use an SD card for small files, as the library to work with those have great overhead scketch size wise and ram (they need a fixed 512 bytes just for the buffer). Use an EEPROM instead and connect to the Wire/I2C interface. I can't because I used the interface pins as GPIOs for buttons, well that's on me.

* trying some things to lower memory footprint

* untested, it just compiles :), refactor

* fix pc version

* WIP removed sd card and added an EEPROM to store the levels, code to take advantage of that, not yet tested to see it working

* minor refactor around the useless logger

* some more tidying up

* disabled enemies, audio and status bar in order to get the code to fit the arduino, long road ahead to get everything inside the flash bounds of arduino..